genericBotBase/requirements.txt
discord.py

asyncio
python-dateutil
humanize
aiohttp